pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
requires-python = ">=3.12.9,<4.0"
dependencies = [
    "fastapi==0.115.12",
    "httpx==0.28.1",
    "uvicorn==0.34.0",
    "loguru==0.7.3",
    "sqlalchemy==2.0.36",
    "asyncpg==0.30.0",
    "pyjwt==2.10.1",
    "passlib==1.7.4",
    "pytest==8.3.5",
    "pytest-asyncio==0.26.0",
    "pytest-cov==6.1.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
exclude = ["**/__init__.py"]
line-length = 100